requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
beautifulsoup4
tqdm